chore: created nestedProperty.ts in utils file and improved createNestedProperty function
